pyproject.toml
[project]
name = "bjj-vqa"
version = "0.1.0"
description = "Brazilian Jiu-Jitsu Visual Q&A Benchmark"
readme = "README.md"
requires-python = ">=3.13"
license = "GPL-3.0-or-later"
keywords = ["bjj", "jiu-jitsu", "vqa", "benchmark", "multimodal", "vision-language"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
]
urls = { Repository = "https://github.com/matheusccouto/bjj-vqa", Dataset = "https://huggingface.co/datasets/couto/bjj-vqa" }
dependencies = [
    "inspect-ai>=0.3.0",
    "openai>=1.0.0",
    "huggingface-hub>=0.20.0",
    "datasets>=2.14.0",
    "pillow>=10.0.0",
    "pydantic>=2.0.0",
]

[project.scripts]
bjj-vqa = "bjj_vqa.cli:main"

[dependency-groups]
dev = [
    {include-group = "test"},
    {include-group = "lint"},
]
test = [
    "pytest>=8.0.0",
]
lint = [
    "ruff>=0.9.0",
    "ty>=0.0.1",
]

[build-system]
requires = ["uv_build>=0.11.3,<0.12"]
build-backend = "uv_build"

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = ["ALL"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "ANN",    # type annotations not required in tests
    "S101",   # assert-used
    "D401",   # docstring imperative mood - fixtures describe what they return
    "PLC0415",# import not at top level - needed for env var setup
    "ARG002", # unused method argument - pytest fixtures injected by framework
]
"src/bjj_vqa/cli.py" = [
    "T201",   # print-used - CLI uses print for output
]
"scripts/**/*.py" = [
    "INP001", # implicit namespace package - scripts are standalone
    "T201",   # print-used - scripts output to stdout
    "PLR2004",# magic value comparison - acceptable for simple scripts
]

[tool.ty.environment]
python = ".venv"
root = ["./src"]

[tool.ty.src]
include = ["src", "tests"]